+++ src/controllers/index.js
@@ -1,6 +1,13 @@
 import Calculator from './Calculator.js';
 import EventController from './EventController.js';
+import Money from './Money.js';
 import OutputController from './OutputController.js';
 import InputController from './InputController.js';
 
-export { Calculator, EventController, InputController, OutputController };
+export {
+  Calculator,
+  EventController,
+  InputController,
+  Money,
+  OutputController,
+};
